½ºÅ͵ð¸ñ·Ï UI¿¡ ´ëÇÑ Ŀ¹Ô
